Made matched filter more selective
Now you can create a smaller matched filter that just cares about a few
pixels of interest. You can assigned this MF to a RefCube object and also
give it a list of the raveled pixel indices that the MF corresponds to.
These will automatically be unraveled when it comes time to apply.
Also, generate_matched_filter can now accept a RefCube object.