RDI MF methods now accept a ReferenceCube object
I decided that a ReferenceCube object will also keep track of the KL
basis generated with that reference cube, and the matched filter
generated with the KL basis. It made sense to modify
generate_matched_filter() and apply_matched_filter_to_image() so that
they can accept a reference cube object.