API changes to klip_subtract_with_basis
Return order is changed. It makes much more sense to have the first axis
index different images (if they are passed to the function), so that the
following axes index the different KL parameters for one particular
image. This has the added advantage that the output can be directly
passed to MatchedFilter.apply_matched_filter_to_images().