Added new functionality to utils, and tests for it
Now, utils has a function that lets you take an image and generate a stamp
centered at each pixel of the image. You can tell it which pixels to choose.
This is helpful because then you can use the dot-product matched filter, which
doesn't have the same frequency issues that the FFT filter does, and it happens
very fast.